Update eec-qa-wo.js
Updated for Actual Finish Date which is a required field for our installation. Because of how the script tries to find the required inputs and how they are using a input from Telerik, the script can't catch it. We have to add a required tag to the blank input and set the value to the input and the script can now count as it was intended.